refactor(b64): make all dependencies optional
Additionally, move the types into logical modules for better organization.
