+++ jose/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [features]
-jwk = []
-jws = ["p256?/ecdsa", "p384?/ecdsa", "dep:serde_json", "jose-b64/serde_json"]
+jwk = ["jose-b64/secret"]
+jws = ["p256?/ecdsa", "p384?/ecdsa", "dep:serde_json", "jose-b64/json"]
 sha2 = ["dep:sha2", "dep:digest"]
 hmac = ["dep:rand_core", "dep:hmac"]
 p256 = ["dep:rand_core", "dep:p256"]
